# bookingservice/api/bookingservice/v1/booking_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             v6.32.0
// source: bookingservice/v1/booking.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBookingServiceCancelBooking = "/booking.v1.BookingService/CancelBooking"
const OperationBookingServiceConfirmBooking = "/booking.v1.BookingService/ConfirmBooking"
const OperationBookingServiceCreateBooking = "/booking.v1.BookingService/CreateBooking"
const OperationBookingServiceGetBookedSeats = "/booking.v1.BookingService/GetBookedSeats"
const OperationBookingServiceGetBooking = "/booking.v1.BookingService/GetBooking"
const OperationBookingServiceGetEvent = "/booking.v1.BookingService/GetEvent"
const OperationBookingServiceGetLockedSeats = "/booking.v1.BookingService/GetLockedSeats"
const OperationBookingServiceListBookings = "/booking.v1.BookingService/ListBookings"
const OperationBookingServiceLockSeat = "/booking.v1.BookingService/LockSeat"
const OperationBookingServiceUnlockSeat = "/booking.v1.BookingService/UnlockSeat"
const OperationBookingServiceUpdateBooking = "/booking.v1.BookingService/UpdateBooking"

type BookingServiceHTTPServer interface {
	CancelBooking(context.Context, *CancelBookingRequest) (*CreateBookingReply, error)
	ConfirmBooking(context.Context, *ConfirmBookingRequest) (*CreateBookingReply, error)
	CreateBooking(context.Context, *CreateBookingRequest) (*CreateBookingReply, error)
	GetBookedSeats(context.Context, *GetBookedSeatsRequest) (*GetBookedSeatsReply, error)
	GetBooking(context.Context, *GetBookingRequest) (*CreateBookingReply, error)
	GetEvent(context.Context, *GetEventRequest) (*GetEventReply, error)
	GetLockedSeats(context.Context, *GetLockedSeatsRequest) (*GetLockedSeatsReply, error)
	ListBookings(context.Context, *ListBookingsRequest) (*ListBookingsReply, error)
	LockSeat(context.Context, *LockSeatRequest) (*LockSeatReply, error)
	UnlockSeat(context.Context, *UnlockSeatRequest) (*UnlockSeatReply, error)
	UpdateBooking(context.Context, *UpdateBookingRequest) (*UpdateBookingReply, error)
}

func RegisterBookingServiceHTTPServer(s *http.Server, srv BookingServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/bookings", _BookingService_CreateBooking0_HTTP_Handler(srv))
	r.GET("/v1/bookings/{id}", _BookingService_GetBooking0_HTTP_Handler(srv))
	r.GET("/v1/bookings", _BookingService_ListBookings0_HTTP_Handler(srv))
	r.PATCH("/v1/bookings/{id}", _BookingService_UpdateBooking0_HTTP_Handler(srv))
	r.PUT("/v1/bookings/{id}", _BookingService_CancelBooking0_HTTP_Handler(srv))
	r.PUT("/v1/bookings/{id}/confirm", _BookingService_ConfirmBooking0_HTTP_Handler(srv))
	r.GET("/events/{event_id}/booked-seats", _BookingService_GetBookedSeats0_HTTP_Handler(srv))
	r.GET("/events/{event_id}/locked-seats", _BookingService_GetLockedSeats0_HTTP_Handler(srv))
	r.POST("/v1/events/{event_id}/lock-seat", _BookingService_LockSeat0_HTTP_Handler(srv))
	r.POST("/v1/events/{event_id}/unlock-seat", _BookingService_UnlockSeat0_HTTP_Handler(srv))
	r.GET("/v1/events/{id}", _BookingService_GetEvent0_HTTP_Handler(srv))
}

func _BookingService_CreateBooking0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateBookingRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceCreateBooking)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateBooking(ctx, req.(*CreateBookingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateBookingReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_GetBooking0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetBookingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceGetBooking)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetBooking(ctx, req.(*GetBookingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateBookingReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_ListBookings0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListBookingsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceListBookings)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListBookings(ctx, req.(*ListBookingsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListBookingsReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_UpdateBooking0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateBookingRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceUpdateBooking)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateBooking(ctx, req.(*UpdateBookingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateBookingReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_CancelBooking0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CancelBookingRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceCancelBooking)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CancelBooking(ctx, req.(*CancelBookingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateBookingReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_ConfirmBooking0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ConfirmBookingRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceConfirmBooking)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ConfirmBooking(ctx, req.(*ConfirmBookingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateBookingReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_GetBookedSeats0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetBookedSeatsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceGetBookedSeats)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetBookedSeats(ctx, req.(*GetBookedSeatsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetBookedSeatsReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_GetLockedSeats0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetLockedSeatsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceGetLockedSeats)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetLockedSeats(ctx, req.(*GetLockedSeatsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetLockedSeatsReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_LockSeat0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LockSeatRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceLockSeat)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LockSeat(ctx, req.(*LockSeatRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LockSeatReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_UnlockSeat0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UnlockSeatRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceUnlockSeat)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UnlockSeat(ctx, req.(*UnlockSeatRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UnlockSeatReply)
		return ctx.Result(200, reply)
	}
}

func _BookingService_GetEvent0_HTTP_Handler(srv BookingServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetEventRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBookingServiceGetEvent)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetEvent(ctx, req.(*GetEventRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetEventReply)
		return ctx.Result(200, reply)
	}
}

type BookingServiceHTTPClient interface {
	CancelBooking(ctx context.Context, req *CancelBookingRequest, opts ...http.CallOption) (rsp *CreateBookingReply, err error)
	ConfirmBooking(ctx context.Context, req *ConfirmBookingRequest, opts ...http.CallOption) (rsp *CreateBookingReply, err error)
	CreateBooking(ctx context.Context, req *CreateBookingRequest, opts ...http.CallOption) (rsp *CreateBookingReply, err error)
	GetBookedSeats(ctx context.Context, req *GetBookedSeatsRequest, opts ...http.CallOption) (rsp *GetBookedSeatsReply, err error)
	GetBooking(ctx context.Context, req *GetBookingRequest, opts ...http.CallOption) (rsp *CreateBookingReply, err error)
	GetEvent(ctx context.Context, req *GetEventRequest, opts ...http.CallOption) (rsp *GetEventReply, err error)
	GetLockedSeats(ctx context.Context, req *GetLockedSeatsRequest, opts ...http.CallOption) (rsp *GetLockedSeatsReply, err error)
	ListBookings(ctx context.Context, req *ListBookingsRequest, opts ...http.CallOption) (rsp *ListBookingsReply, err error)
	LockSeat(ctx context.Context, req *LockSeatRequest, opts ...http.CallOption) (rsp *LockSeatReply, err error)
	UnlockSeat(ctx context.Context, req *UnlockSeatRequest, opts ...http.CallOption) (rsp *UnlockSeatReply, err error)
	UpdateBooking(ctx context.Context, req *UpdateBookingRequest, opts ...http.CallOption) (rsp *UpdateBookingReply, err error)
}

type BookingServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBookingServiceHTTPClient(client *http.Client) BookingServiceHTTPClient {
	return &BookingServiceHTTPClientImpl{client}
}

func (c *BookingServiceHTTPClientImpl) CancelBooking(ctx context.Context, in *CancelBookingRequest, opts ...http.CallOption) (*CreateBookingReply, error) {
	var out CreateBookingReply
	pattern := "/v1/bookings/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBookingServiceCancelBooking))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) ConfirmBooking(ctx context.Context, in *ConfirmBookingRequest, opts ...http.CallOption) (*CreateBookingReply, error) {
	var out CreateBookingReply
	pattern := "/v1/bookings/{id}/confirm"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBookingServiceConfirmBooking))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) CreateBooking(ctx context.Context, in *CreateBookingRequest, opts ...http.CallOption) (*CreateBookingReply, error) {
	var out CreateBookingReply
	pattern := "/v1/bookings"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBookingServiceCreateBooking))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) GetBookedSeats(ctx context.Context, in *GetBookedSeatsRequest, opts ...http.CallOption) (*GetBookedSeatsReply, error) {
	var out GetBookedSeatsReply
	pattern := "/events/{event_id}/booked-seats"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBookingServiceGetBookedSeats))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) GetBooking(ctx context.Context, in *GetBookingRequest, opts ...http.CallOption) (*CreateBookingReply, error) {
	var out CreateBookingReply
	pattern := "/v1/bookings/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBookingServiceGetBooking))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) GetEvent(ctx context.Context, in *GetEventRequest, opts ...http.CallOption) (*GetEventReply, error) {
	var out GetEventReply
	pattern := "/v1/events/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBookingServiceGetEvent))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) GetLockedSeats(ctx context.Context, in *GetLockedSeatsRequest, opts ...http.CallOption) (*GetLockedSeatsReply, error) {
	var out GetLockedSeatsReply
	pattern := "/events/{event_id}/locked-seats"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBookingServiceGetLockedSeats))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) ListBookings(ctx context.Context, in *ListBookingsRequest, opts ...http.CallOption) (*ListBookingsReply, error) {
	var out ListBookingsReply
	pattern := "/v1/bookings"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBookingServiceListBookings))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) LockSeat(ctx context.Context, in *LockSeatRequest, opts ...http.CallOption) (*LockSeatReply, error) {
	var out LockSeatReply
	pattern := "/v1/events/{event_id}/lock-seat"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBookingServiceLockSeat))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) UnlockSeat(ctx context.Context, in *UnlockSeatRequest, opts ...http.CallOption) (*UnlockSeatReply, error) {
	var out UnlockSeatReply
	pattern := "/v1/events/{event_id}/unlock-seat"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBookingServiceUnlockSeat))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BookingServiceHTTPClientImpl) UpdateBooking(ctx context.Context, in *UpdateBookingRequest, opts ...http.CallOption) (*UpdateBookingReply, error) {
	var out UpdateBookingReply
	pattern := "/v1/bookings/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBookingServiceUpdateBooking))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}